requirements.txt
streamlit==1.45.1
tensorflow==2.10.0
numpy==1.24.4
Pillow==9.5.0